+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+
+Adds support for Conviva Analytics with Bitmovin Player React Native SDK.
+
+### Added
+
+- `ConvivaAnalytics` class and related types to enable Conviva Analytics with Bitmovin Player React Native SDK.
+- Simple React Native app under `example/` to showcase the integration.